backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
groq==0.8.0
langchain==0.2.1
langchain-groq==0.1.3
langgraph==0.0.62
langchain-community==0.2.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
xgboost==2.0.3
scikit-learn==1.4.2
shap==0.45.0
pandas==2.2.2
numpy==1.26.4
vaderSentiment==3.3.2
transformers==4.41.1
torch==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
python-multipart==0.0.9
aiosqlite==0.20.0
